refactor(client): move client event emitter to constants.ts
The client event emitter was previously defined in index.ts, which was also the entry point for the client module. To improve modularity and separation of concerns, the event emitter has been moved to constants.ts. This change also required updating all imports that previously referenced the event emitter from index.ts to now reference constants.ts. The index.ts file has been removed as it is no longer needed.